update: 타입 클래스를 별도의 패키지로 분리
