+++ src/main/java/com/newdeal/ledger/cardaccount/dto/AccountRequest.java
@@ -1,5 +1,7 @@
 package com.newdeal.ledger.cardaccount.dto;
 
+import com.newdeal.ledger.cardaccount.dto.type.AccountType;
+
 import lombok.Data;
 
 public class AccountRequest {
