--- src/main/java/com/newdeal/ledger/cardaccount/dto/AccountType.java
@@ -1,5 +1,0 @@
-package com.newdeal.ledger.cardaccount.dto;
-
-public enum AccountType {
-	MONEY, ACCOUNT
-}
