--- src/main/java/com/newdeal/ledger/cardaccount/dto/CardType.java
@@ -1,5 +1,0 @@
-package com.newdeal.ledger.cardaccount.dto;
-
-public enum CardType {
-	CREDIT, CHECK
-}
