+++ src/main/java/com/newdeal/ledger/categorytag/dto/type/CategoryType.java
@@ -1,4 +1,4 @@
-package com.newdeal.ledger.categorytag.dto;
+package com.newdeal.ledger.categorytag.dto.type;
 
 public enum CategoryType {
 	INCOME, EXPENSE;
