+++ src/main/java/com/newdeal/ledger/transaction/dto/SourceDto.java
@@ -1,5 +1,8 @@
 package com.newdeal.ledger.transaction.dto;
 
+import com.newdeal.ledger.transaction.dto.type.SourceType;
+import com.newdeal.ledger.transaction.dto.type.SubSourceType;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
