--- src/main/java/com/newdeal/ledger/transaction/dto/SourceType.java
@@ -1,5 +1,0 @@
-package com.newdeal.ledger.transaction.dto;
-
-public enum SourceType {
-	ACCOUNT, CARD;
-}
