+++ src/main/java/com/newdeal/ledger/transaction/dto/type/SourceType.java
@@ -1,0 +1,5 @@
+package com.newdeal.ledger.transaction.dto.type;
+
+public enum SourceType {
+	ACCOUNT, CARD;
+}
