+++ src/main/java/com/newdeal/ledger/transaction/dto/type/SubSourceType.java
@@ -1,4 +1,4 @@
-package com.newdeal.ledger.transaction.dto;
+package com.newdeal.ledger.transaction.dto.type;
 
 public enum SubSourceType {
 	MONEY, ACCOUNT, CREDIT, CHECK
